fix: 타입 변경
