+++ src/modals/SelectTeamModal.tsx
@@ -1,6 +1,6 @@
 import { Modal } from '@/components/Modal';
 import { css } from '@/styled-system/css';
-import { Team } from '@/types';
+import { Team } from '@/typesOld';
 
 export type SelectTeamModalProps = {
   isOpen: boolean;
