+++ src/pages/Room/Entry.tsx
@@ -4,7 +4,7 @@ import { Button } from '@/components/Button';
 import { mockTeams } from '@/mock/data';
 import { css } from '@/styled-system/css';
 import { container, vstack } from '@/styled-system/patterns';
-import { Team } from '@/types';
+import { Team } from '@/typesOld';
 
 type EntryProps = {
   setRole: (role: 'admin' | 'player') => void;
