+++ src/utils/user.ts
@@ -1,4 +1,4 @@
-import { User } from '@/types';
+import { User } from '@/typesOld';
 
 export const generateId = () => {
   return Math.random().toString(36).substr(2, 9);
